+++ src/main/scala/satify/update/solver/dpll/utils/DpllUtils.scala
@@ -1,9 +1,9 @@
-package satify.update.solver.dpll
+package satify.update.solver.dpll.utils
 
 import satify.model.CNF
 import satify.model.dpll.{Decision, PartialModel}
 import satify.update.solver.dpll.Dpll.dpll
-import satify.update.solver.dpll.PartialModelUtils.{explodeSolutions, extractModelFromCnf, extractSolutionsFromDT}
+import satify.update.solver.dpll.utils.PartialModelUtils.{explodeSolutions, extractModelFromCnf, extractSolutionsFromDT}
 
 object DpllUtils:
 
@@ -12,3 +12,5 @@ object DpllUtils:
       for pmSet <- extractSolutionsFromDT(dpll(Decision(extractModelFromCnf(cnf), cnf)))
       yield explodeSolutions(pmSet)
     s.flatten
+    
+
