+++ src/main/scala/satify/update/solver/dpll/utils/PartialModelUtils.scala
@@ -1,4 +1,4 @@
-package satify.update.solver.dpll
+package satify.update.solver.dpll.utils
 
 import satify.model.Bool.True
 import satify.model.CNF.*
@@ -7,9 +7,7 @@ import satify.model.dpll.OrderedSeq.*
 import satify.model.dpll.{Constraint, Decision, DecisionTree, PartialModel}
 import satify.model.{CNF, Variable}
 
-private object PartialModelUtils:
-
-  import satify.model.dpll.OrderedSeq.given_Ordering_Variable
+object PartialModelUtils:
 
   import satify.model.dpll.OrderedSeq.given_Ordering_Variable
