+++ src/test/scala/satify/update/solver/dpll/PartialModelUtilsTest.scala
@@ -7,7 +7,7 @@ import satify.model.dpll.OrderedSeq.{given_Ordering_Variable, seq}
 import satify.model.dpll.{Constraint, Decision, PartialModel}
 import satify.model.{CNF, Variable}
 import satify.update.solver.dpll.Dpll.dpll
-import satify.update.solver.dpll.PartialModelUtils.*
+import satify.update.solver.dpll.utils.PartialModelUtils.*
 
 class PartialModelUtilsTest extends AnyFlatSpec with Matchers:
